modif classes, functions, graphics, constantes
